End of the day push to remote
