+++ server.js
@@ -2,7 +2,6 @@ const express = require('express')
 const cors = require('cors')
 const dotenv = require('dotenv')
 const mongoose = require('./providers/mongoDB')
-const redisClient = require('./providers/redisClient')
 const morgan = require('morgan')
 const cookie_parser = require('cookie-parser')
 const app = express()
